requirements.txt
streamlit
pandas
torch
transformers
spacy
sentence-transformers
faiss-cpu
plotly-express
scikit-learn